added track list display

+++ app/src/main/java/com/evg/deezerplayer/navigation/Route.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 sealed interface Route {
     @Serializable data object Chart: Route
     @Serializable data object DownloadedTracks: Route
-    @Serializable data class TrackPlayer(val id: Int): Route
+    @Serializable data class TrackPlayer(val id: Long): Route
 }
